target-app/requirements.txt
fastapi
uvicorn
paho-mqtt
prometheus-client